Resolving clojure-lsp problem with same namespaces on different paths
In the case of same namespaces on different paths in project (in pasta where user ns in test and dev folders, and app.core in backend and frontend) on my local machines one of these files is ignored by clojure-lsp formatting and clean-ns.  May machines are bot Mac (intel and M1) and looks like it is not the problem on GitHub??
I resolved problem by renaming test/user to test/test-repl, frontend.app/core to frontend.app/core-fe and backend.app/core to  backend.app/core-be.