Replacing funicular var to function
Sometimes after restarting system in the moment of reloading user-repl ns dev-sys atom is still nil and var funicular is defined as nil (specially if restart is initiated with backend watcher).
To prevent this behavior we are using function instead of var that is evaluated in the moment of running command and dev-sys is guaranteed to be populated.